Fixed a couple rendering bugs I introduced, and improve loading perf. (#742)
- I left a debugging string ("Chevron") in the text, which
would show up on screen. Oops.

- The trick I used to glue the footer to the bottom of the browser window ended up causing trouble with the
vertical scroll bar positioning. I reverted this trick, in favor of a simpler approach. Instead of gluing
the footer to the bottom of the window, I simply make it so the bottom of the page is already rendered in the
same blue as the footer.

- Move a script from loading in the header to loading in the body. According to the Chrome profiler, this
saves a bunch of time blocking rendering during page load.