Switch uses of <figure> to {% include figure.html %} (#818)
figure.html now contains a bit of magic to substantially improve
the load time and visual appearance of pages that contain imagery.
The magic pre-allocates space on the page of the exact size needed
for the image, even though the image isn't loaded yet. The result
is that the page can be layed out once, layout doesn't need to
happen again for each image that gets loaded.