Fix logo hover state rotation (#2781)
* Fix logo hover state rotation

The logo rotates upon hover but it was not centered correctly. This was caused
because no explicit height was defined for the logo container. By doing so
we now can have the transform use 50%,50% as the center point for smooth
rotations.

This change has been tested in Chrome Version 69.0.3497.100 (Official Build)

* Make the logo be actually circulair

The SVG radius did not account for the border of the circle
causing it to be chopped off at the edges. By subtracting
the border size from the radius (*2) we now have a perfect
looking circle.

This change has been tested in Chrome Version 69.0.3497.100 (Official Build)

* make build scss => css