Refine the multicluster gateways installation documentation (#4307)
* Refine the multicluster gateways installation documentation

Change all occurances of we to you.
Use the kubernetes version macro instead of hard-coding.
Refinements to the wording in various places.
Improvements to the workflow, so it works 100% of the time instead of most of the time.
Add the a warning flag related to the samples certs.
Add more information about stubbing a domain for cloud provides that have specific unique procedures.

* refinements

* Fix shortcode error

* Fix linting problems

* Apply review comments.