Add Istio Deployment Models concept. (#4813)
* Add Istio Deployment Models concept.

This concept replaces the old multi-cluster concept.
Includes new diagrams that comply with the diagram creation guidelines.
Updates the Chinese content to use a local copy of the previous diagrams.
Fixes all internal links to the previous version of the doc.

Signed-off-by: rcaballeromx <grca@google.com>

* Add glossary entries for needed terms.

The terms involved are:

    - Cluster
    - Identity
    - Trust domain

Signed-off-by: rcaballeromx <grca@google.com>

* Define cluster in a platform agnostic way.

Also adds links between `identity` and `trust domain`.

Signed-off-by: rcaballeromx <grca@google.com>

* Add missing `(` in links.

Signed-off-by: rcaballeromx <grca@google.com>

* Fix links to sections and reduce image sizes.

Signed-off-by: rcaballeromx <grca@google.com>

* Simplify the definition of `trust domain`

Signed-off-by: rcaballeromx <grca@google.com>

* Move old images to the ZH content.

Signed-off-by: rcaballeromx <grca@google.com>

* Add reworked control plane content.

Also addresses the comments left on the PR including those regarding the
diagrams.

Signed-off-by: rcaballeromx <grca@google.com>

* Add fail over example and glossary entries.

This update also reworks the control plane models section to fit the example.
Additional adjustments were made to the diagrams too.

Signed-off-by: rcaballeromx <grca@google.com>

* Move mesh models section.

Also minor fixes and edits.

Signed-off-by: rcaballeromx <grca@google.com>

* Fix glossary entries and links.

Signed-off-by: rcaballeromx <grca@google.com>