Fix TCP Traffic Shifting test script (#7016)
The `TestTCPTrafficShifting` test script relies on fetching the port
named "tcp" on the istio-ingressgateway service, but that port is not
present in the default installation profile -- only in the demo
profile.  This patches the service to add the port in the setup
script, and also fixes the usage of `nc` in the test script.