Fix flaky circuit breaker test (#7043)
The `TestCircuitBreaker` test is very flaky.  It relies on examining
the percentage of requests that succeed or fail to a service with
circuit breaking configured, and ensuring the numbers fall within a
given range.  Unfortunately, the percentages tend to vary widely from
run to run, so the test fails as often as it passes.

This simplifies the test by only verifying that both successes and
failures occur instead of checking the percentage of each.