Remove DOC_OWNERS.md (#8504)
This has proven to cause constant merge conflicts and decrease
contribution experience. We went through the same issue in the
environments WG and got rid of these types of generated codes which just
lead to constant conflicts.

If we do really want this, we should move it to something generated on
the fly in the build and not checked in.