No longer need this warning about Kiali addon install. (#10081)
Now that the Kiali addon has been upgraded to v1.36, there is no longer the monitoring dashboard CRD that we have to worry about. This is what caused that timing error (the CRD would fail to be established in time before the dashboards themselves started to get created).

Since this timing error won't happen, we can remove this warning in the docs.